+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - add: automatically load `app_path/data/db-ui` as the last metadata source when it exists
 - add: support `_delete` metadata tombstones for removing entities before export
 - add: support `!` metadata clear instructions for scalar and relation fields
+- add: support `!id` metadata relation removal instructions
 
 ## 0.23.1 (2026-05-07)